Introducing new branch for testing other solutions:
In this commit, a new branch of the Train Dispatch System application has been created. Theres also been added some new classes, as well as refactoring the package and class structures for a more robust application. The other focus is on testing a new direction by introducing iterators, streams, and lambda functions to see if these changes lead to a better solution for the application. Additionally, the test classes have also been enhanced, and introduced more test for the essential methods and classes.